Refactor a bit and add missing annotations
